Support fake hipSolver
