+++ cupy/cuda/cupy_cuda.h
@@ -5,18 +5,13 @@
 
 #include <stdint.h>
 
-#if !defined(CUPY_NO_CUDA) && !defined(CUPY_USE_HIP)
-#include <cuda.h>
-
-#endif  // #if !defined(CUPY_NO_CUDA) && !defined(CUPY_USE_HIP)
-
-
 #if CUPY_USE_HIP
 
 #include "cupy_hip.h"
 
 #elif !defined(CUPY_NO_CUDA)
 
+#include <cuda.h>
 #include <cublas_v2.h>
 #include <cuda_profiler_api.h>
 #include <cuda_runtime.h>
@@ -51,7 +46,8 @@ cublasStatus_t cublasGetMathMode(...) {
 
 #else // #ifndef CUPY_NO_CUDA
 
-#define CUDA_VERSION 0
+#include "cupy_cuda_common.h"
+#include "cupy_cuComplex.h"
 
 extern "C" {
 
@@ -60,31 +56,6 @@ bool hip_environment = false;
 ///////////////////////////////////////////////////////////////////////////////
 // cuda.h
 ///////////////////////////////////////////////////////////////////////////////
-
-typedef int CUdevice;
-typedef enum {
-    CUDA_SUCCESS = 0,
-} CUresult;
-enum CUjit_option {};
-enum CUjitInputType {};
-enum CUfunction_attribute {};
-
-
-typedef void* CUdeviceptr;
-struct CUctx_st;
-struct CUevent_st;
-struct CUfunc_st;
-struct CUmod_st;
-struct CUstream_st;
-struct CUlinkState_st;
-
-
-typedef struct CUctx_st* CUcontext;
-typedef struct CUevent_st* cudaEvent_t;
-typedef struct CUfunc_st* CUfunction;
-typedef struct CUmod_st* CUmodule;
-typedef struct CUstream_st* cudaStream_t;
-typedef struct CUlinkState_st* CUlinkState;
 
 // Error handling
 CUresult cuGetErrorName(...) {
@@ -171,35 +142,6 @@ CUresult cuFuncSetAttribute(...) {
 ///////////////////////////////////////////////////////////////////////////////
 // cuda_runtime.h
 ///////////////////////////////////////////////////////////////////////////////
-
-enum {
-    cudaDevAttrComputeCapabilityMajor = 75,
-    cudaDevAttrComputeCapabilityMinor = 76,
-};
-
-typedef enum {
-    cudaSuccess = 0,
-    cudaErrorInvalidValue = 1,
-    cudaErrorMemoryAllocation = 2,
-} cudaError_t;
-typedef enum {} cudaDataType;
-enum cudaDeviceAttr {};
-enum cudaMemoryAdvise {};
-enum cudaMemcpyKind {};
-
-
-typedef void (*cudaStreamCallback_t)(
-    cudaStream_t stream, cudaError_t status, void* userData);
-
-
-struct cudaPointerAttributes{
-    int device;
-    void* devicePointer;
-    void* hostPointer;
-    int isManaged;
-    int memoryType;
-};
-
 
 // Error handling
 const char* cudaGetErrorName(...) {
@@ -383,29 +325,10 @@ cudaError_t cudaEventSynchronize(...) {
     return cudaSuccess;
 }
 
-///////////////////////////////////////////////////////////////////////////////
-// cuComplex.h
-///////////////////////////////////////////////////////////////////////////////
-
-#include "cupy_cuComplex.h"
 
 ///////////////////////////////////////////////////////////////////////////////
 // cublas_v2.h
 ///////////////////////////////////////////////////////////////////////////////
-
-typedef void* cublasHandle_t;
-
-typedef enum {} cublasDiagType_t;
-typedef enum {} cublasFillMode_t;
-typedef enum {} cublasOperation_t;
-typedef enum {} cublasPointerMode_t;
-typedef enum {} cublasSideMode_t;
-typedef enum {} cublasGemmAlgo_t;
-typedef enum {} cublasMath_t;
-typedef enum {
-    CUBLAS_STATUS_SUCCESS=0,
-} cublasStatus_t;
-
 
 // Context
 cublasStatus_t cublasCreate(...) {
