+++ cupy/cuda/cupy_cuda_common.h
@@ -1,0 +1,93 @@
+// This file is a stub header file of cuda for Read the Docs.
+
+#ifndef INCLUDE_GUARD_CUPY_CUDA_COMMON_H
+#define INCLUDE_GUARD_CUPY_CUDA_COMMON_H
+
+#define CUDA_VERSION 0
+
+extern "C" {
+
+///////////////////////////////////////////////////////////////////////////////
+// cuda.h
+///////////////////////////////////////////////////////////////////////////////
+
+typedef int CUdevice;
+typedef enum {
+    CUDA_SUCCESS = 0,
+} CUresult;
+enum CUjit_option {};
+enum CUjitInputType {};
+enum CUfunction_attribute {};
+
+
+typedef void* CUdeviceptr;
+struct CUctx_st;
+struct CUevent_st;
+struct CUfunc_st;
+struct CUmod_st;
+struct CUstream_st;
+struct CUlinkState_st;
+
+
+typedef struct CUctx_st* CUcontext;
+typedef struct CUevent_st* cudaEvent_t;
+typedef struct CUfunc_st* CUfunction;
+typedef struct CUmod_st* CUmodule;
+typedef struct CUstream_st* cudaStream_t;
+typedef struct CUlinkState_st* CUlinkState;
+
+
+///////////////////////////////////////////////////////////////////////////////
+// cuda_runtime.h
+///////////////////////////////////////////////////////////////////////////////
+
+enum {
+    cudaDevAttrComputeCapabilityMajor = 75,
+    cudaDevAttrComputeCapabilityMinor = 76,
+};
+
+typedef enum {
+    cudaSuccess = 0,
+    cudaErrorInvalidValue = 1,
+    cudaErrorMemoryAllocation = 2,
+} cudaError_t;
+typedef enum {} cudaDataType;
+enum cudaDeviceAttr {};
+enum cudaMemoryAdvise {};
+enum cudaMemcpyKind {};
+
+
+typedef void (*cudaStreamCallback_t)(
+    cudaStream_t stream, cudaError_t status, void* userData);
+
+
+struct cudaPointerAttributes{
+    int device;
+    void* devicePointer;
+    void* hostPointer;
+    int isManaged;
+    int memoryType;
+};
+
+
+///////////////////////////////////////////////////////////////////////////////
+// cublas_v2.h
+///////////////////////////////////////////////////////////////////////////////
+
+typedef void* cublasHandle_t;
+
+typedef enum {} cublasDiagType_t;
+typedef enum {} cublasFillMode_t;
+typedef enum {} cublasOperation_t;
+typedef enum {} cublasPointerMode_t;
+typedef enum {} cublasSideMode_t;
+typedef enum {} cublasGemmAlgo_t;
+typedef enum {} cublasMath_t;
+typedef enum {
+    CUBLAS_STATUS_SUCCESS=0,
+} cublasStatus_t;
+
+} // extern "C"
+
+#endif // #ifndef CUPY_NO_CUDA
+#endif // #ifndef INCLUDE_GUARD_CUPY_CUDA_COMMON_H
