+++ cupy/cuda/cupy_cusolver.h
@@ -3,14 +3,18 @@
 #ifndef INCLUDE_GUARD_CUPY_CUSOLVER_H
 #define INCLUDE_GUARD_CUPY_CUSOLVER_H
 
-#include "cupy_cuda.h"
-
-#ifndef CUPY_NO_CUDA
+#if !defined(CUPY_NO_CUDA) && !defined(CUPY_USE_HIP)
 
 #include <cusolverDn.h>
 #include <cusolverSp.h>
 
-#else // #ifndef CUPY_NO_CUDA
+#else // #if !defined(CUPY_NO_CUDA) && !defined(CUPY_USE_HIP)
+
+#ifdef CUPY_USE_HIP
+#include "cupy_hip_common.h"
+#else // #ifdef CUPY_USE_HIP
+#include "cupy_cuda_common.h"
+#endif // #ifdef CUPY_USE_HIP
 
 extern "C" {
 
@@ -275,5 +279,5 @@ cusolverStatus_t cusolverSpDcsrlsvchol(...) {
 
 } // extern "C"
 
-#endif // #ifndef CUPY_NO_CUDA
+#endif // #if !defined(CUPY_NO_CUDA) && !defined(CUPY_USE_HIP)
 #endif // #ifndef INCLUDE_GUARD_CUPY_CUSOLVER_H
